requirements.txt
flask
protobuf==3.17.3
google-api-python-client==1.12.11
https://cdn.mysql.com/Downloads/Connector-Python/mysql-connector-python-2.0.4.tar.gz
google-auth-httplib2 
google-auth-oauthlib 
